data analysis - titanic dataset
Data Analysis
Data analysis is the process of cleaning, transforming, and modeling data to extract useful information and draw conclusions from it. It is an essential part of data science and plays a significant role in decision-making processes across various industries. In this README file, we will cover the basics of data analysis using popular Python libraries such as Pandas, NumPy, and Matplotlib.

Setting up the Environment
To get started with data analysis in Python, we need to set up the environment by installing the necessary packages. The easiest way to do this is to use Anaconda, which is a popular Python distribution that comes pre-packaged with all the essential data science libraries.

Alternatively, we can also use Google Colab, a cloud-based platform for data analysis, that provides access to pre-installed packages and an interface for coding in Python.